Refactor validator
- validation 을 위한 조건, 설정 등에 관한 변수 validationConfig 생성
- 여러가지 입력값 검증에 통합적으로 사용할 수 있는 함수 checkValidity 생성
- 가입 요청 전송 전 모든 입력값이 조건에 부합하는지 검사하기 위한 객체 isMandatoryArgsValid 생성
- js 파일간 변수명 중복 방지를 위해 즉시 실행 함수 형식으로 변경
