+++ static/js/register.js
@@ -1,6 +1,7 @@
-import "./validators.js";
+import "./validator.js";
 import "./terms.js";
 import "./optionalInfo.js";
 import "./map.js";
 import "./telAuth.js";
 import "./submit.js";
+import "./isInputsValid.js";
