+++ views/register.pug
@@ -10,13 +10,13 @@ block content
     p 필수 정보를 입력해주세요.
     form
       .id
-        input(type="text" placeholder="아이디* (4자 이상)")
+        input(type="text" name="id" placeholder="아이디* (4자 이상)")
         span
       .pw
-        input(type="password" placeholder="비밀번호* (8자 이상)")
+        input(type="password" name="pw" placeholder="비밀번호* (8자 이상)")
         span
       .pw2
-        input(type="password" placeholder="비밀번호 재확인*")
+        input(type="password" name="pw2" placeholder="비밀번호 재확인*")
         span
       .row.mail
         input(type="text" placeholder="이메일 앞자리*")
@@ -26,11 +26,11 @@ block content
         each val in ["naver.com", "hanmail.net", "nate.com", "hotmail.com", "gmail.com", "직접입력"]
           option= val
       .name
-        input(type="text" placeholder="이름*")
+        input(type="text" name="name" placeholder="이름*")
         span
       .tel-auth
         .row.tel-auth__req
-          input(type="tel" placeholder="휴대폰*")
+          input(type="tel" name="tel" placeholder="휴대폰*")
           button.deactive(disabled) 인증받기
         span
       ul
